requirements.txt
streamlit
numpy
pillow
tensorflow-cpu 
matplotlib
opencv-python-headless